+++ ViewportDrivenGraphStreaming/src/main/java/aljoschaRydzyk/viewportDrivenGraphStreaming/FlinkOperator/Vertex/VertexFilterMaxDegree.java
@@ -4,14 +4,14 @@ import org.apache.flink.api.common.functions.FilterFunction;
 import org.apache.flink.types.Row;
 
 public class VertexFilterMaxDegree implements FilterFunction<Row>{
-	private Integer numberVertices;
+	private long numberVertices;
 	
-	public VertexFilterMaxDegree(Integer numberVertices) {
+	public VertexFilterMaxDegree(long numberVertices) {
 		this.numberVertices = numberVertices;
 	}
 	@Override
 	public boolean filter(Row value) throws Exception {
-		if ((Integer) value.getField(2) < numberVertices) return true;
+		if ((long) value.getField(2) < numberVertices) return true;
 		else return false;
 	}
 }
